Moved Demo select to separate file
